% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{missing_summary}
\alias{missing_summary}
\title{Generate a table that summarizes patterns of missingness.}
\usage{
missing_summary(data, vars = NULL, show.non.missing = TRUE)
}
\arguments{
\item{data}{Data object, of class "data.frame" or "tibble".}

\item{vars}{Character vector, if not NULL, identifies which variables should be included in the table.}

\item{show.non.missing}{Logical object, if not TRUE, table will only display variables with missing values.}
}
\value{
A gt table object that summarizes patterns of missingness.
}
\description{
Given a data set, create a table that numerically summarizes aspects of missing data.
}
