% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{mlr_table}
\alias{mlr_table}
\title{Generate MLR results table}
\usage{
mlr_table(model, var_names)
}
\arguments{
\item{model}{The MLR model whose results should be used}

\item{var_names}{A named list with entries of the format parameter = "Parameter Label"}
}
\value{
A gt table object
}
\description{
Given a multivariate linear regression model and the labels for parameters, create a results table
}
\examples{

# create MLR model

m1 <- lm(mpg ~ wt + am)

# make named list of parameter labels

parameter_names <- list(wt = "Weight", am = "Transmission")

# generate results table

mlr_table(model = m1, var_names = parameter_names)

}
