% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{Generate diagnostic plots for linear regression model}
\usage{
diagnostic_plots(
  model,
  line_color = "steelblue",
  line_size = 1,
  point_color = "blue",
  point_shape = 1,
  point_alpha = 0.6
)
}
\arguments{
\item{model}{The linear regression model to create diagnostic plots for}

\item{line_color}{The color of lines in the plots. Default is "steelblue"}

\item{line_size}{The size of lines in the plots. Default is 1}

\item{point_color}{The color of outliers in the plots. Default is "blue"}

\item{point_shape}{The shape of points in the plots. Default is 1}

\item{point_alpha}{The alpha of points in the plots. Default is 0.6}
}
\value{
A list of ggplot plot objects in the following order: Residual vs Fit, QQ Plot, Scale-Location, Leverage Plot, grid of all plots
}
\description{
Given a linear regression model and aesthetic parameters, generate plots to check assumptions of linear regression
}
\examples{

# create linear regression model

data(mtcars)

m1 <- lm(mpg ~ wt + am, data = mtcars)

# generate diagnostic plots

diagnostic_plots(model = m1)

}
